usb: pd: policy_engine: somc_ext: Cast return value to silence warning
The return value is checked for IS_ERR_VALUE, which expects a long.

+++ drivers/usb/pd/policy_engine.c
@@ -1257,7 +1257,7 @@ static void phy_wait_vbus_settled_down(struct usbpd *pd, int mv,
 		msleep(itvl);
 		rc = power_supply_get_property(pd->usb_psy,
 				POWER_SUPPLY_PROP_VOLTAGE_NOW, &pval);
-		if (IS_ERR_VALUE(rc))
+		if (IS_ERR_VALUE((long)rc))
 			goto waitremain;
 		usbin = pval.intval / 1000;
 		cnt++;
